pyproject.toml
[tool.poetry]
name = "pip-ai"
version = "0.1.0"
description = "pip.ai is a Currency Exchange Monitor & Alert Agent. It is a Python-based tool that allows users to track and receive alerts on currency exchange rate fluctuations. This project uses the uAgent library to create a user-friendly interface and to interact with a currency exchange API to provide real-time exchange rate data."
authors = [
    "Vedant Tamhane <vedant.tamhane03@gmail.com>", 
    "Gagankumar C Kummur <gckummur@gmail.com>", 
    "Eva Ghanshani <evaghanshani@gmail.com>",
    "Sukhpreet Singh <sssukhpreet2003@gmail.com>"
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
aiohttp = "3.8.5"
aiosignal = "1.3.1"
altair = "5.1.2"
apispec = "6.3.0"
async-timeout = "4.0.3"
attrs = "23.1.0"
bech32 = "1.2.0"
blinker = "1.6.3"
cachetools = "5.3.1"
certifi = "2023.7.22"
charset-normalizer = "3.3.0"
click = "8.1.7"
colorama = "0.4.6"
cosmpy = "0.9.1"
ecdsa = "0.18.0"
everapi = "0.1.1"
frozenlist = "1.4.0"
gitdb = "4.0.10"
gitpython = "3.1.37"
googleapis-common-protos = "1.60.0"
grpcio = "1.51.1"
h11 = "0.14.0"
idna = "3.4"
importlib-metadata = "6.8.0"
jinja2 = "3.1.2"
jsonschema = "4.19.1"
jsonschema-specifications = "2023.7.1"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
mdurl = "0.1.2"
msgpack = "1.0.7"
multidict = "6.0.4"
numpy = "1.26.0"
packaging = "23.2"
pandas = "2.1.1"
pillow = "10.0.1"
protobuf = "4.24.3"
pyarrow = "13.0.0"
pycryptodome = "3.19.0"
pydantic = "1.10.13"
pydeck = "0.8.1b0"
pygments = "2.16.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
referencing = "0.30.2"
requests = "2.31.0"
rich = "13.6.0"
rpds-py = "0.10.3"
six = "1.16.0"
smmap = "5.0.1"
streamlit = "1.27.2"
tenacity = "8.2.3"
toml = "0.10.2"
toolz = "0.12.0"
tornado = "6.3.3"
typing-extensions = "4.8.0"
tzdata = "2023.3"
tzlocal = "5.1"
uagents = "0.6.2"
urllib3 = "2.0.6"
uvicorn = "0.20.0"
validators = "0.22.0"
watchdog = "3.0.0"
websockets = "10.4"
yarl = "1.9.2"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
